Release 5.11.0 (#288)
* Improve retry and statistic calculation performance

* Update statistic calculation triggers

* Fix deadlock

* Cluster entity provided. Log table updated with reference

* Formatting fix

* Cascade removing fix

* References removed

* Not null constraint provided

* Project id reference provided

* Search attribute added

* Added onboarding

* Clusters table structure fix

* EPMRPP-68409 || cluster item impl (#206)

EPMRPP-68409 M2M cluster_test_item table added

* EPMRPP-68775 || Auto unique error attributes provided (#209)

* EPMRPP-68656/EPMRPP-68659 || Fix onboarding problems order and naming (#211)

* Update 55_onboarding_fix.up.sql

* EPMRPP-65502 || Scripts renamed

* EPMRPP-68947 || Add alter ticket script (#213)

* EPMRPP-68947 || Fix version number

* EPMRPP-68649 || 54 unique errors attributes down scripts fix (#212)

* EPMRPP-73239 || Stale materialized view table added (#216)

* Added SSLMODE variable

* Docker-compose updataed

* Docker-compose updataed

* Changed to POSTGRES_SSLMODE

* EPMRPP-75872 || user_creation_bid fields extension

* 5.7.0 || Merge master to develop (#222)

* Added Force SSL parameter (#218)


* 5.7.0 || Merge 5.7 to master (#221)

* Improve retry and statistic calculation performance

* Update statistic calculation triggers

* Fix deadlock

* Cluster entity provided. Log table updated with reference

* Formatting fix

* Cascade removing fix

* References removed

* Not null constraint provided

* Project id reference provided

* Search attribute added

* Added onboarding

* Clusters table structure fix

* EPMRPP-68409 || cluster item impl (#206)

EPMRPP-68409 M2M cluster_test_item table added

* EPMRPP-68775 || Auto unique error attributes provided (#209)

* EPMRPP-68656/EPMRPP-68659 || Fix onboarding problems order and naming (#211)

* Update 55_onboarding_fix.up.sql

* EPMRPP-65502 || Scripts renamed

* EPMRPP-68947 || Add alter ticket script (#213)

* EPMRPP-68947 || Fix version number

* EPMRPP-68649 || 54 unique errors attributes down scripts fix (#212)

* EPMRPP-73239 || Stale materialized view table added (#216)

* Added SSLMODE variable

* Docker-compose updataed

* Docker-compose updataed

* Changed to POSTGRES_SSLMODE



* EPMRPP-72301 || Add Organization table to DB

* EPMRPP-72301 || Typo fixed

* EPMRPP-77644 || add rule_name field to sender_case + migration

* EPMRPP-77644 || changed type from text to varchar

* EPMRPP-77644 || updated length of rule_name field & marked as not null

* Newjenkins (#226)

* Update Jenkinsfile

* Update Jenkinsfile

* Update Jenkinsfile

* Update Jenkinsfile

* Update Jenkinsfile

* Update Jenkinsfile

* RC pipeline updated

* tagging updated

* spaces

* EPMRPP-78986 || Updated issue type hex colors

* EPMRPP-78986 || Added down script

* EPMRPP-78986 || Updated wrong field in WHERE clause

* EPMRPP-79337 || Updated colors for sub issue types (#235)

EPMRPP-79337 || Updated colors for sub issue types

* Fix missed Postgres custom port (#215)

Fixes https://github.com/reportportal/migrations/issues/210

* EPMRPP-84251 || Bump Script version

* EPMRPP-84251 || Bump Script version

* EPMRPP-84316-update-activity-table added migration script

* EPMRPP-84316-update-activity-table added cascade by project

* EPMRPP-84114 || Add last used field for api_key (#251)

* EPMRPP-84316-update-activity-table added minor changes

* Revert "EPMRPP-84316 || Update activity table"

* EPMRPP-78914 || Redesign integrations. Link updates (#272)

* EPMRPP-78914 || Redesign integrations. Link updates

* added github action rc build images

* changed exit-code

* EPMRPP-85814 || Remove redundant index on test item table (#273)

* Merge setting branch

* Fix typo

* Delete duplicate

* EPMRPP-84226 || Remove access token logic from the code

* Merge master to develop

* EPMRPP-86891 fixed broken links in Help&Support modal (#277)

* Add GitHub action for building dev image

* Delete .github/workflows/rc.yaml

* Refactor build image workflows (#281)

* EPMRPP-85814 || Fix down script

* EPMRPP-87666 || Drop gin index on trgm log message (#284)

* EPMRPP-89168 || Add project attribute interval cast

---------

+++ migrations/18_job_attributes.up.sql
@@ -1,5 +1,5 @@
 UPDATE project_attribute
-SET value = extract(epoch FROM value::interval)
+SET value = extract(epoch FROM value::interval)::BIGINT
 WHERE attribute_id IN (SELECT id
                        FROM attribute
                        WHERE attribute.name IN ('job.interruptJobTime',
